# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Post Your Opinion API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/"
        }
    ],
    "paths": {
        "/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get list of blogs",
                "description": "Returns list of blogs",
                "operationId": "get Blogs ",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "blogs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "User Id",
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "description": "Blog Title",
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "description": "Blog Content",
                                                        "type": "string"
                                                    },
                                                    "created_by": {
                                                        "description": "User Id of the blog created",
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "description": "Blog Creation timestamp",
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "description": "Blog Updated timestamp",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        },
                                        "message": {
                                            "description": "Message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create new blog",
                "description": "Create new blog",
                "operationId": "create blog",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Blog Title",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Blog Content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "User Id",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Blog Title",
                                            "type": "string"
                                        },
                                        "content": {
                                            "description": "Blog Content",
                                            "type": "string"
                                        },
                                        "created_by": {
                                            "description": "User Id of the blog created",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Blog Creation timestamp",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Blog Updated timestamp",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}